More fixes. Fix more!

+++ src/core/stats/rank/Staff.php
@@ -29,7 +29,7 @@ class Staff extends Rank {
     	return self::STAFF;
     }
 
-    public function getChatTime() : int {
+    public function getChatTime() : float {
     	return 0;
     }
 }
